qcacld-3.0: Return failure for runtime resume if FW is down
Currently when the FW is down, runtime resume return success
status but does not send the power exit and WOW wake up from
sleep command to FW. So WMI_WOW_SET_ACTION_WAKE_UP_CMDID
command is sent to FW which leads to panic.

To avoid this issue, send the correct error status for runtime
resume if FW is down.

CRs-Fixed: 2830614
Change-Id: Id27d2648427b32d32c15370fbb170fafec9450d2
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>