qcacld-3.0: Update he and ht rate flags correctly
Currently driver does not update he and ht rate flags correctly
in wma as rate flags should include all the subsets of the
lower rate sets, which is not the case today and driver only
updates the higher rate flag. Because of which it leads to
invalid computation of txrate at the kernel.

To address this issue, update he and ht rate flags correctly in
wma.

change-Id: I6ec68af1f75650990384efbd61536042329665b2
CRs-Fixed: 2837245
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>