msm: camera: isp: Change csid stop sequence during flush
This change takes care of a race which we see when we halt master
csid immediately and slave csid halt to follow. In some scenarios
after master csid halt, slave gives violation errors.
With this change we are first changing the halt mode of slave to
internal so that dependency on master is removed and
then issue halt commands to both master and slave.

CRs-Fixed: 2830502
Change-Id: I05fd1d07298c7dddf3d5024e0c0cacdf15e2f69d
Signed-off-by: Chandan Kumar Jha <cjha@codeaurora.org>