qcacld-3.0: Debug the reason for recovery during driver unload
During the driver unload whenever the recovery is triggered in
the current implementation the request is silently dropped, this
is resulting issues in other areas during the driver which is making
it harder to debug the issue.

Trigger system panic to debug the reason for recovery during the
driver unload.

Change-Id: I67c6c7e8b19dcc3371cd6691c0e3f19be4c33c46
CRs-Fixed: 2850591
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>