qcacld-3.0: Don't block NAN+NDI+STA concurrency for non-dbs
When a new connection is about to come up, host checks if current
concurrency combination including the new connection is allowed or
not based on the HW capability.
Firmware manages NAN + NDI by dividing up slots. Connection on NDI
is re-negotiable and therefore a 3rd connection with the same MAC
is possible.

Change-Id: I63e39c5cd4945cd308e475c1e03f676336c4e7c1
CRs-Fixed: 2841457
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>