msm: camera: isp: Add dynamic enabling of IRQ in controller
CSID SOFs are needed to do a health check of sensors once the
sensor freeze is observed. Since the 680 CSID driver has controller
based mechanism, there is a need to enable the IRQ on controller
dynamically.
This commit adds support for dynamic enable/disable of IRQ on
controller.
Minor clean ups to remove unused code.

CRs-Fixed: 2830502
Change-Id: Ic4985555a5726264606a8e60c725eb8d1ce81364
Signed-off-by: Gaurav Jindal <gjindal@codeaurora.org>