qca-wifi: Do not add 80MHz non DFS channels in precaclist
While initializing the DFS preCAC lists in a SBS device, 36HT80
is added a valid tree node and preCAC is triggered on the channel.
Do not add 80MHz non DFS channels in the preCAC list and handle
the restricted 80p80 as a special case.

Add a new API in regulatory to find out if a given frequency
and width combination has any DFS sub frequencies in it.

Change-Id: I4ca3abbec63e303e48ca9c271c90ce542d3d12ab
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>