qcacld-3.0: Update multi-ko support for hermetic builds
The current multi-ko support is achieved by creating, during the
actual build process, target-specific .<chip> folders beneath the
qcacld-3.0 folder, and then populating those folders with a variety of
symbolic links. Unfortunately this mechanism violates the requirements
of the Hermetic Build feature which requires that the source directory
not be modified during the build process.

One solution to this problem is to actually add the target-specific
folders, along with the included symlinks, to the qcacld-3.0 git
project. But one maintenance issue with this approach is that these
folders of symlinks would require update when the qcacld-3.0 directory
structure changes.

An alternative solution is that instead of having each .<chip> file be
a folder of symlinks, let each .<chip> file be symbolic link. This
approach reduces the maintenance issue since you only need to add a
new .<chip> symbolic link when a new target is added; no changes are
required to the symbolic links when a new qcacld-3.0 folder is added.

As a precursor to transitioning to this alternate solution, update the
current build mechanism such that only .<chip> symbolic links are
created in the source directory, and update the remainder of the
multi-ko logic to align with the resulting change in pathnames.

A subsequent change can remove the dynamic symbolic link creation, and
instead add the symbolic links to the git project.

Change-Id: Ie3df49217ffec260085285b1ae25113c3a943e9b
CRs-Fixed: 2889020
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>